+++ packages/webblocks-cms/resources/lang/de/blocks.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+  'search_form' => [
+    'label' => 'Suche',
+    'placeholder' => 'Diese Website durchsuchen',
+    'submit' => 'Suchen',
+  ],
+];