+++ packages/webblocks-cms/resources/lang/en/blocks.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+  'search_form' => [
+    'label' => 'Search',
+    'placeholder' => 'Search this site',
+    'submit' => 'Search',
+  ],
+];