+++ packages/webblocks-cms/resources/lang/tr/blocks.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+  'search_form' => [
+    'label' => 'Arama',
+    'placeholder' => 'Bu sitede ara',
+    'submit' => 'Ara',
+  ],
+];